change data type of lockIn duration

+++ pojo/src/main/java/com/stanzaliving/core/payment/dto/SoldBookingDto.java
@@ -15,7 +15,7 @@ public class SoldBookingDto {
     String roomUuid;
     Date contractStartDate;
     Date lockInEndDate;
-    Double lockInDuration;
+    Integer lockInDuration;
     Double beds;
 
     List<String> packageService;
